Added classes
Added classes to contain roads and objects

The road class can contain the road name, type and a list of object classes it is connected to. This list is in the correct order, starting from either end (as specified in the comment).

The object class is more sophisticated. It contains the object ID on the map and what roads it is connected to, as you would expect. It has one method, getinstruction, which returns a user-friendly instruction from getting from the specified "In road" to the specified "Out road".